requirements.txt
pybit==2.5.0
pandas==2.0.0
numpy==1.24.3
python-dotenv==1.0.0
ccxt==3.1.54
ta==0.10.2
requests==2.30.0
schedule==1.2.0
